refactor: 악보 entity 및 악보 메뉴 sql 문 일부 수정 - mapper.xml에서는 select를 전체가 아닌 출력할 부분만 설정 - entity 클래스는 sno를 Long 형으로 변경 -> 자료가 많으면 Integer는 제한적임

+++ clnewze-back/src/main/java/com/clnewze/back/clnewzeback/domain/entity/SheetMusic.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class SheetMusic {
-  private Integer sno;
+  private Long sno;
   private String artist;
   private String title;
   private String img;
